made the prepared statement manager even better
